+++ azalea-protocol/src/packets/game/c_container_set_data.rs
@@ -3,7 +3,8 @@ use azalea_protocol_macros::ClientboundGamePacket;
 
 #[derive(Clone, Debug, AzBuf, ClientboundGamePacket)]
 pub struct ClientboundContainerSetData {
-    pub container_id: i8,
+    #[var]
+    pub container_id: i32,
     pub id: u16,
     pub value: u16,
 }
