+++ azalea-protocol/src/packets/game/c_container_set_slot.rs
@@ -4,7 +4,8 @@ use azalea_protocol_macros::ClientboundGamePacket;
 
 #[derive(Clone, Debug, AzBuf, ClientboundGamePacket)]
 pub struct ClientboundContainerSetSlot {
-    pub container_id: i8,
+    #[var]
+    pub container_id: i32,
     #[var]
     pub state_id: u32,
     pub slot: u16,
