+++ azalea-protocol/src/packets/game/s_container_click.rs
@@ -6,7 +6,8 @@ use azalea_protocol_macros::ServerboundGamePacket;
 
 #[derive(Clone, Debug, AzBuf, ServerboundGamePacket)]
 pub struct ServerboundContainerClick {
-    pub container_id: u8,
+    #[var]
+    pub container_id: i32,
     #[var]
     pub state_id: u32,
     pub slot_num: i16,
