+++ azalea-protocol/src/packets/game/s_container_close.rs
@@ -3,5 +3,6 @@ use azalea_protocol_macros::ServerboundGamePacket;
 
 #[derive(Clone, Debug, AzBuf, ServerboundGamePacket)]
 pub struct ServerboundContainerClose {
-    pub container_id: u8,
+    #[var]
+    pub container_id: i32,
 }
