+++ azalea-protocol/src/packets/game/s_place_recipe.rs
@@ -4,7 +4,8 @@ use azalea_protocol_macros::ServerboundGamePacket;
 
 #[derive(Clone, Debug, AzBuf, ServerboundGamePacket)]
 pub struct ServerboundPlaceRecipe {
-    pub container_id: u8,
+    #[var]
+    pub container_id: i32,
     pub recipe: ResourceLocation,
     pub shift_down: bool,
 }
